split up conf into multiple files
